+++ samples/service-account/TokenGenerationWithContextExample.ts
@@ -15,12 +15,15 @@ let cred = {
     privateKey: "<YOUR_PEM_privateKey>",
 };
 
-function getSkyflowBearerTokenFromFilePath() {
+function getSkyflowBearerTokenWithContextFromFilePath() {
     return new Promise(async (resolve, reject) => {
         try {
+            const options = {
+                context: "context_id",
+            };
             if (!isExpired(bearerToken)) resolve(bearerToken);
             else {
-                let response = await generateBearerToken(filepath);
+                let response = await generateBearerToken(filepath, options);
                 bearerToken = response.accessToken;
                 resolve(bearerToken);
             }
@@ -30,13 +33,17 @@ function getSkyflowBearerTokenFromFilePath() {
     });
 }
 
-function getSkyflowBearerTokenFromCreds() {
+function getSkyflowBearerTokenWithContextFromCreds() {
     return new Promise(async (resolve, reject) => {
         try {
+            const options = {
+                context: "context_id",
+            };
             if (!isExpired(bearerToken)) resolve(bearerToken);
             else {
                 let response = await generateBearerTokenFromCreds(
-                    JSON.stringify(cred)
+                    JSON.stringify(cred),
+                    options
                 );
                 bearerToken = response.accessToken;
                 resolve(bearerToken);
@@ -48,8 +55,8 @@ function getSkyflowBearerTokenFromCreds() {
 }
 
 const tokens = async () => {
-    console.log(await getSkyflowBearerTokenFromFilePath());
-    console.log(await getSkyflowBearerTokenFromCreds());
+    console.log(await getSkyflowBearerTokenWithContextFromFilePath());
+    console.log(await getSkyflowBearerTokenWithContextFromCreds());
 
 };